reuse unique_name logic
