+++ trimesh/exchange/gltf.py
@@ -18,6 +18,7 @@ from .. import rendering
 from .. import resources
 from .. import transformations
 
+from ..util import unique_name
 from ..caching import hash_fast
 from ..constants import log, tol
 
@@ -378,6 +379,7 @@ def load_glb(file_obj,
                            ignore_broken=ignore_broken,
                            merge_primitives=merge_primitives,
                            mesh_kwargs=mesh_kwargs)
+
     return kwargs
 
 
@@ -802,12 +804,13 @@ def _append_mesh(mesh,
         (include_normals is None and
          'vertex_normals' in mesh._cache.cache)):
         # store vertex normals if requested
-        #
-        normals = mesh.vertex_normals.copy()
         if unitize_normals:
+            normals = mesh.vertex_normals.copy()
             norms = np.linalg.norm(normals, axis=1)
             if not util.allclose(norms, 1.0, atol=1e-4):
                 normals /= norms.reshape((-1, 1))
+            else:
+                normals = mesh.vertex_normals
 
         acc_norm = _data_append(
             acc=tree['accessors'],
@@ -1208,49 +1211,6 @@ def _parse_materials(header, views, resolver=None):
             materials.append(visual.material.PBRMaterial(**pbr))
 
     return materials
-
-
-def unique_name(start, contains):
-    """
-    Deterministically generate a unique name not
-    contained in a dict. Will create names of the
-    form "start_10" and increment accordingly.
-
-    Parameters
-    -----------
-    start : str
-      Initial guess for name
-    contains : dict, set, or list
-      Bundle of existing names we cannot use.
-
-    Returns
-    ---------
-    unique : str
-      A name that is not contained in `contains`
-    """
-    # exit early if name is not in bundle
-    if len(contains) == 0 or (len(start) > 0 and start not in contains):
-        return start
-
-    increment = 0
-    formatter = start + '_{}'
-    if len(start) > 0:
-        # split by our delimiter once
-        split = start.rsplit('_', 1)
-        if len(split) == 2 and split[1].isnumeric():
-            if split[0] not in contains:
-                return split[0]
-            # start incrementing from the passed value
-            increment = int(split[1])
-            formatter = split[0] + '_{}'
-
-    # if contains is empty we will only need to check once
-    for i in range(increment + 1, 2 + increment + len(contains)):
-        check = formatter.format(i)
-        if check not in contains:
-            return check
-
-    raise ValueError('unable to establish unique name!')
 
 
 def _read_buffers(header,
