+++ trimesh/scene/scene.py
@@ -1,20 +1,20 @@
 import numpy as np
 import collections
 
+from . import cameras
+from . import lighting
+
 from .. import util
 from .. import units
 from .. import convex
 from .. import caching
 from .. import grouping
 from .. import transformations
-
 from .. import bounds as bounds_module
 
+from ..util import unique_name
 from ..exchange import export
 from ..parent import Geometry3D
-
-from . import cameras
-from . import lighting
 
 from .transforms import SceneGraph
 
@@ -170,9 +170,8 @@ class Scene(Geometry3D):
             # try to create a simple name
             name = 'geometry_' + str(len(self.geometry))
 
-        # if its already taken add a unique random string to it
-        if name in self.geometry:
-            name += ':' + util.unique_id().upper()
+        # if its already taken use our unique name logic
+        name = unique_name(start=name, contains=self.geometry.keys())
 
         # save the geometry reference
         self.geometry[name] = geometry
@@ -180,15 +179,12 @@ class Scene(Geometry3D):
         # create a unique node name if not passed
         if node_name is None:
             # if the name of the geometry is also a transform node
-            if name in self.graph.nodes:
-                # a random unique identifier
-                unique = util.unique_id()
-                # geometry name + UUID
-                node_name = name + '_' + unique.upper()
-                assert node_name not in self.graph.nodes
-            else:
-                # otherwise make the transform node name the same as the geom
-                node_name = name
+            # which graph nodes already exist
+            existing = self.graph.transforms.node_data.keys()
+            # find a name that isn't contained already starting
+            # at the name we have
+            node_name = unique_name(name, existing)
+            assert node_name not in existing
 
         if transform is None:
             # create an identity transform from parent_node
@@ -1213,7 +1209,8 @@ def append_scenes(iterable, common=['world'], base_frame='world'):
         # if a node is consumed and isn't one of the nodes
         # we're going to hold common between scenes remap it
         if node not in common and node in consumed:
-            name = str(node) + '-' + util.unique_id().upper()
+            # generate a name not in consumed
+            name = unique_name(str(node), consumed)
             map_node[node] = name
             node = name
 
@@ -1236,10 +1233,7 @@ def append_scenes(iterable, common=['world'], base_frame='world'):
         map_geom = {}
         for k, v in s.geometry.items():
             # if a geometry already exists add a UUID to the name
-            if k in geometry:
-                name = str(k) + '-' + util.unique_id().upper()
-            else:
-                name = k
+            name = unique_name(start=k, contains=geometry.keys())
             # store name mapping
             map_geom[k] = name
             # store geometry with new name
